+++ packages/health_connector_core/test/src/models/health_data_types/health_data_type_test.dart
@@ -15,6 +15,7 @@ void main() {
             <HealthDataType>[
               HealthDataType.activeEnergyBurned,
               HealthDataType.activityIntensity,
+              HealthDataType.appleStandHour,
               HealthDataType.exerciseTime,
               HealthDataType.moveTime,
               HealthDataType.standTime,