+++ packages/health_connector_hk_ios/ios/health_connector_hk_ios/Sources/health_connector_hk_ios/mappers/PermissionMapper.swift
@@ -125,6 +125,7 @@ extension HealthDataPermissionRequestDto {
              .infrequentMenstrualCycleEvent,
              .highHeartRateEvent,
              .walkingSteadinessEvent,
+             .appleStandHour,
              .persistentIntermenstrualBleedingEvent,
              .prolongedMenstrualPeriodEvent,
              .prolongedMenstrualPeriodEvent,