+++ packages/health_connector_hk_ios/ios/health_connector_hk_ios/Sources/health_connector_hk_ios/mappers/health_record_mappers/HealthRecordMapper.swift
@@ -354,6 +354,8 @@ extension HKCategorySample {
             try toHeadphoneAudioExposureEventRecordDto()
         case .walkingSteadinessEvent:
             try toWalkingSteadinessEventRecordDto()
+        case .appleStandHour:
+            try toAppleStandHourRecordDto()
         default:
             throw HealthConnectorError.invalidArgument(
                 message: