+++ packages/health_connector_hk_ios/ios/health_connector_hk_ios/Sources/health_connector_hk_ios/utils/HealthKitExtensions.swift
@@ -224,6 +224,8 @@ extension HKSample {
                 return .highHeartRateEvent
             case HKCategoryTypeIdentifier.appleWalkingSteadinessEvent.rawValue:
                 return .walkingSteadinessEvent
+            case HKCategoryTypeIdentifier.appleStandHour.rawValue:
+                return .appleStandHour
             // Correlation types
             case HKCorrelationTypeIdentifier.bloodPressure.rawValue:
                 return .bloodPressure